# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Clinical Trial Data Collection Platform API
  description: |
    HIPAA-compliant RESTful API for clinical trial patient data collection platform.
    
    This API provides secure endpoints for:
    - User authentication with JWT and MFA support
    - Patient data management with column-level encryption
    - Questionnaire creation and assignment
    - Secure response collection and storage
    - Media file uploads with virus scanning
    - Comprehensive audit logging
    
    ## Security Features
    - JWT Bearer token authentication
    - Role-based access control (Physician vs Patient)
    - Multi-factor authentication (MFA) support
    - Column-level data encryption for PHI
    - HIPAA-compliant audit logging
    - Rate limiting and request throttling
    - Virus scanning for file uploads
    
    ## Data Privacy
    - All personally identifiable information (PII) is encrypted at rest
    - PHI is never exposed in error messages
    - Comprehensive audit trails for all data access
    - Secure token-based session management
    
    ## Authentication Flow
    1. Login with email/password to receive access token
    2. Include Bearer token in Authorization header for all requests
    3. Refresh tokens before expiration using /auth/refresh
    4. Complete MFA verification if enabled
    
  version: 1.0.0
  contact:
    name: Clinical Trial Platform Development Team
    email: dev@clinicaltrial.platform
  license:
    name: Proprietary
...
servers:
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token obtained from /auth/login endpoint.
        
        Example: Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
  schemas:
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: dr.smith@centralmedical.com
        password:
          type: string
          format: password
          minLength: 8
          description: User's password
          example: DevPassword123!
        userType:
          type: string
          enum: [physician, patient]
          description: Type of user account
          example: physician
        mfaCode:
          type: string
          minLength: 6
          maxLength: 6
          description: MFA code (required if MFA is enabled)
          example: "123456"
      required:
        - email
        - password
        - userType
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Login successful
        user:
          $ref: '#/components/schemas/User'
        accessToken:
          type: string
          description: JWT access token for API authentication
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        expiresIn:
          type: string
          description: Token expiration time
          example: 30m
        requiresMfa:
          type: boolean
          description: Indicates if MFA code is required
          example: false
      required:
        - success
        - message
...
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: patient@example.com
        password:
          type: string
          format: password
          minLength: 8
          description: Secure password meeting complexity requirements
          example: SecurePass123!
        invitationToken:
          type: string
          minLength: 32
          description: Valid invitation token received via email
          example: a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6q7r8s9t0u1v2w3x4y5z6
        userType:
          type: string
          enum: [physician, patient]
          description: Type of user account to create
          example: patient
      required:
        - email
        - password
        - invitationToken
        - userType
...
    ApiError:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
          example: Authentication required
        code:
          type: string
          description: Machine-readable error code
          example: AUTH_REQUIRED
        details:
          type: array
          items:
            type: object
          description: Additional error details (validation errors, etc.)
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
        requestId:
          type: string
          description: Unique request identifier for debugging
      required:
        - error
        - code
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User authentication
      description: |
        Authenticate a user with email and password. Returns JWT access token for subsequent API requests.
        
        If MFA is enabled, this endpoint will return `requiresMfa: true` and the client must call this 
        endpoint again with the `mfaCode` parameter.
        
        Rate limited to 5 attempts per 15 minutes per IP address.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              physician_login:
                summary: Physician login
                value:
                  email: dr.smith@centralmedical.com
                  password: DevPassword123!
                  userType: physician
              patient_login:
                summary: Patient login
                value:
                  email: patient@example.com
                  password: PatientPass123!
                  userType: patient
              mfa_login:
                summary: Login with MFA code
                value:
                  email: dr.smith@centralmedical.com
                  password: DevPassword123!
                  userType: physician
                  mfaCode: "123456"
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              examples:
                successful_login:
                  summary: Successful authentication
                  value:
                    success: true
                    message: Login successful
                    user:
                      id: a1b2c3d4-e5f6-g7h8-i9j0-k1l2m3n4o5p6
                      email: dr.smith@centralmedical.com
                      userType: physician
                      status: active
                      mfaEnabled: false
                      profileCompleted: true
                      createdAt: "2024-01-15T10:30:00Z"
                    accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    expiresIn: "30m"
                mfa_required:
                  summary: MFA code required
                  value:
                    success: true
                    message: MFA code required
                    requiresMfa: true
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                invalid_credentials:
                  summary: Invalid credentials
                  value:
                    error: Invalid credentials
                    code: INVALID_CREDENTIALS
                    timestamp: "2024-01-15T10:30:00Z"
                    requestId: req_a1b2c3d4e5f6g7h8
                invalid_mfa:
                  summary: Invalid MFA code
                  value:
                    error: Invalid MFA code
                    code: INVALID_MFA
                    timestamp: "2024-01-15T10:30:00Z"
                    requestId: req_a1b2c3d4e5f6g7h8
        '423':
          description: Account locked due to failed attempts
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiError'
                  - type: object
                    properties:
                      lockedUntil:
                        type: string
                        format: date-time
                        description: When the account will be unlocked
              example:
                error: Account temporarily locked due to failed login attempts
                code: ACCOUNT_LOCKED
                lockedUntil: "2024-01-15T11:00:00Z"
                timestamp: "2024-01-15T10:30:00Z"
                requestId: req_a1b2c3d4e5f6g7h8
        '429':
          description: Too many authentication attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                error: Too many authentication attempts, please try again later
                code: AUTH_RATE_LIMIT_EXCEEDED
                timestamp: "2024-01-15T10:30:00Z"
                requestId: req_a1b2c3d4e5f6g7h8
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: User registration (invitation-based)
      description: |
        Register a new user account using a valid invitation token.
        
        Registration is invitation-only for security and compliance:
        - Physicians are invited by organization administrators
        - Patients are invited by their assigned physicians
        
        The invitation token must be valid and not expired.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              patient_registration:
                summary: Patient registration
                value:
                  email: patient@example.com
                  password: SecurePass123!
                  invitationToken: a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6q7r8s9t0u1v2w3x4y5z6
                  userType: patient
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Registration successful
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      email:
                        type: string
                        format: email
                      userType:
                        type: string
                        enum: [physician, patient]
        '400':
          description: Registration failed - invalid invitation or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                invalid_invitation:
                  summary: Invalid invitation token
                  value:
                    error: Invalid or expired invitation token
                    code: INVALID_INVITATION
                    timestamp: "2024-01-15T10:30:00Z"
                    requestId: req_a1b2c3d4e5f6g7h8
                validation_error:
                  summary: Validation failed
                  value:
                    error: Validation failed
                    code: VALIDATION_ERROR
                    details:
                      - field: password
                        message: Password must be at least 8 characters
                      - field: email
                        message: Invalid email format
                    timestamp: "2024-01-15T10:30:00Z"
                    requestId: req_a1b2c3d4e5f6g7h8
...
  /physicians/profile:
    get:
      tags:
        - Physicians
      summary: Get physician profile
      description: |
        Retrieve the complete profile information for the authenticated physician.
        
        Includes professional details, organization affiliation, and account settings.
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  profile:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      email:
                        type: string
                        format: email
                      firstName:
                        type: string
                        example: John
                      lastName:
                        type: string
                        example: Smith
                      title:
                        type: string
                        example: Dr.
                      specialization:
                        type: string
                        example: Cardiology
                      licenseNumber:
                        type: string
                        example: MD123456
                      phone:
                        type: string
                        example: +1-555-0123
                      department:
                        type: string
                        example: Internal Medicine
                      role:
                        type: string
                        example: attending_physician
                      permissions:
                        type: array
                        items:
                          type: string
                        example: ["create_questionnaires", "review_responses"]
                      organization:
                        type: object
                        properties:
                          name:
                            type: string
                            example: Central Medical Center
                          type:
                            type: string
                            example: hospital
    put:
      tags:
        - Physicians
      summary: Update physician profile
      description: |
        Update the physician's profile information.
        
        Only the provided fields will be updated. Sensitive fields like email and 
        license number may require additional verification.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                  minLength: 1
                  maxLength: 50
                lastName:
                  type: string
                  minLength: 1
                  maxLength: 50
                title:
                  type: string
                  maxLength: 20
                specialization:
                  type: string
                  maxLength: 100
                licenseNumber:
                  type: string
                  maxLength: 50
                phone:
                  type: string
                  pattern: '^\+?[1-9]\d{1,14}$'
                department:
                  type: string
                  maxLength: 100
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Profile updated successfully
                  profile:
                    type: object
                    description: Updated profile data
...
  /patients/profile:
    get:
      tags:
        - Patients
      summary: Get patient profile
      description: |
        Retrieve the complete profile information for the authenticated patient.
        
        Includes personal information (decrypted for the patient), medical history,
        and account settings. All PII is encrypted at rest and decrypted for display.
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  profile:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      email:
                        type: string
                        format: email
                      patientId:
                        type: string
                        example: PT-2024-001
                      firstName:
                        type: string
                        example: Jane
                      lastName:
                        type: string
                        example: Doe
                      phone:
                        type: string
                        example: +1-555-0456
                      address:
                        type: string
                        example: 123 Main St, City, ST 12345
                      emergencyContact:
                        type: string
                        example: John Doe - Spouse - +1-555-0457
                      gender:
                        type: string
                        enum: [male, female, other, prefer_not_to_say]
                      status:
                        type: string
                        enum: [active, inactive, completed, withdrawn]
                      enrollmentDate:
                        type: string
                        format: date
                      consentSigned:
                        type: boolean
                      consentDate:
                        type: string
                        format: date-time
                      consentVersion:
                        type: string
                        example: "2.1"
                      mfaEnabled:
                        type: boolean
                      createdAt:
                        type: string
                        format: date-time
    put:
      tags:
        - Patients
      summary: Update patient profile
      description: |
        Update the patient's profile information.
        
        All PII fields are automatically encrypted before storage. Medical history
        and sensitive data require appropriate validation and audit logging.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                  minLength: 1
                  maxLength: 50
                lastName:
                  type: string
                  minLength: 1
                  maxLength: 50
                phone:
                  type: string
                  pattern: '^\+?[1-9]\d{1,14}$'
                address:
                  type: string
                  maxLength: 200
                emergencyContact:
                  type: string
                  maxLength: 200
                gender:
                  type: string
                  enum: [male, female, other, prefer_not_to_say]
                medicalHistory:
                  type: string
                  maxLength: 2000
                medications:
                  type: string
                  maxLength: 1000
                allergies:
                  type: string
                  maxLength: 1000
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Profile updated successfully
                  updatedAt:
                    type: string
                    format: date-time
...
  /media/upload:
    post:
      tags:
        - Media
      summary: Upload single file
      description: |
        Upload a single media file (image, video, audio, or document).
        
        Security features:
        - File type validation against allowed MIME types
        - Virus scanning (results available via separate endpoint)
        - File size limits (100MB maximum)
        - Automatic file hash generation for integrity checking
        - Secure random filename generation
        
        Supported file types:
        - Images: JPEG, PNG, GIF, WebP
        - Videos: MP4, MPEG, QuickTime, AVI
        - Audio: MP3, WAV, OGG, M4A
        - Documents: PDF, TXT, DOC, DOCX
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: File to upload
                responseId:
                  type: string
                  format: uuid
                  description: Associate file with a response (optional)
                tags:
                  type: string
                  description: Comma-separated tags for categorization
                  example: "symptom,rash,day-7"
                description:
                  type: string
                  maxLength: 500
                  description: File description
              required:
                - file
      responses:
        '201':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: File uploaded successfully
                  file:
                    $ref: '#/components/schemas/MediaFile'
        '400':
          description: Upload validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                no_file:
                  summary: No file provided
                  value:
                    error: No file uploaded
                    code: FILE_REQUIRED
                invalid_type:
                  summary: Invalid file type
                  value:
                    error: File type application/executable not allowed
                    code: INVALID_FILE_TYPE
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                error: File size exceeds 100MB limit
                code: FILE_TOO_LARGE
...